Mitigate for missing activesupport blank method
So it happens than when we removed the activesupport dependency
the blank? method in the message (usually a hash) was not longer
available (causing the condition to fail).

The reason why this was not captured in the test is that the test
were testing a payload being a string (which seems to have a blank
method implemented somewhere in logstash).

The new test verifies that if the payload is a hash (which it usually
is), then no unhandled exceptions should happen (happy path).